+++ changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 All changes made after 05/30/2020 will be documented in this file.
 
+## [06/05/2021]
+## Added
+- saving on/off switch state to storage and updating switch state when the page loads
+
 ## [06/04/2021]
 ## Added
 - cleaned up `defaultPopup.html` 
@@ -182,7 +186,8 @@ All changes made after 05/30/2020 will be documented in this file.
 ## [Before]
 Changes before 05/30/2020 are not logged in this changelog, but you can press [before] to see the commit history.
 
-[06/03/2021]: https://github.com/Gbillington1/OneLine-Chrome-Extension/compare/94c4392..HEAD
+[06/05/2021]: https://github.com/Gbillington1/OneLine-Chrome-Extension/compare/0f3bc2e..HEAD
+[06/04/2021]: https://github.com/Gbillington1/OneLine-Chrome-Extension/compare/94c4392..0f3bc2e
 [06/03/2021]: https://github.com/Gbillington1/OneLine-Chrome-Extension/compare/ea7847f..c289ee1
 [05/31/2021]: https://github.com/Gbillington1/OneLine-Chrome-Extension/compare/7afa9de..ecc46b4
 [05/03/2021]: https://github.com/Gbillington1/OneLine-Chrome-Extension/compare/94dd99e..ce24e12